Fix sal issues

+++ sal/linux/sal_sound.c
@@ -7,14 +7,14 @@
 // 48000 Hz maximum; 1/50 of a second; 3 frames to hold (2 plus a bit extra)
 #define BUFFER_SAMPLES (48000 / 50 * (BUFFER_FRAMES + 1))
 
-extern void S9xMixSamples (u8 *buffer, int sample_count);
+extern void S9xMixSamples (u16 *buffer, int sample_count);
 
 static SDL_AudioSpec audiospec;
 
 volatile static unsigned int ReadPos, WritePos;
 
 // 2 channels per sample (stereo); 2 bytes per sample-channel (16-bit)
-static uint8_t Buffer[BUFFER_SAMPLES * 2 * 2];
+static u16 Buffer[BUFFER_SAMPLES * 2 * 2];
 static u32 SamplesPerFrame, BytesPerSample;
 static u32 Muted; // S9xSetAudioMute(TRUE) gets undone after SNES Global Mute ends
